+++ src/ambition/version_details.cc
@@ -41,6 +41,7 @@ std::vector<VersionUpdates> versionDetails = {
         .items = {
           _("Preferentially use idle builders outside of buildings further away instead of from closer buildings."),
           _("Moved the \"Done\" text to the right in the War Factory and Harbour build menus, so that they are not covered up by the \"Esc\" keybind display."),
+          _("Use lower case for default save filenames, based on the king's name."),
         },
       },
     },